fix images path
