+++ package.json
@@ -3,14 +3,17 @@
   "private": true,
   "scripts": {
     "bootstrap": "yarn && node init.js && lerna clean && yarn && lerna bootstrap",
-    "build": "lerna run build",
+    "build": "yarn build:client && yarn build:server",
+    "build:client": "lerna run build --scope=client",
+    "build:server": "lerna run build --scope=server",
     "dev:client": "lerna run dev  --scope=client",
     "dev:server": "lerna run dev --scope=server",
     "dev": "lerna run dev",
     "test": "lerna run test",
     "lint": "lerna run lint",
     "format": "lerna run format",
-    "preview": "lerna run preview"
+    "preview": "lerna run preview",
+    "start": "yarn build && lerna run preview --scope=server"
   },
   "license": "MIT",
   "workspaces": [
