+++ packages/client/public/images/parking-card.png
