+++ packages/client/public/images/start-card.png
