+++ packages/client/src/pages/home/home.tsx
@@ -20,19 +20,29 @@ export const Home = (props: IHomeProps) => {
 
   return (
     <main className={classNames(style.main, isDarkTheme && style.dark)}>
-      <div className={style.mainContent}>
-        <h1 className={style.title}>{title}</h1>
-        <p className={style.description}>{description}</p>
-        <Link className={style.link} to={links.setup.path}>
-          <Button
-            className={style.button}
-            variation={ButtonVariation.PRIMARY}
-            size={ButtonSizes.LARGE}
-            rounded>
-            {linkText}
-          </Button>
-        </Link>
-      </div>
+      {!user && (
+        <>
+          <div className={style.mainContent}>
+            <h1 className={style.title}>{title}</h1>
+            <p className={style.description}>{description}</p>
+            <Link className={style.link} to={links.game.path}>
+              <Button
+                className={style.button}
+                variation={ButtonVariation.PRIMARY}
+                size={ButtonSizes.LARGE}
+                rounded>
+                {linkText}
+              </Button>
+            </Link>
+          </div>
+          <img
+            className={style.cardsImage}
+            src="/images/cards.svg"
+            alt="monopoly cards"
+          />
+        </>
+      )}
+      {user && <></>}
     </main>
   );
 };
