+++ packages/client/vite.config.ts
@@ -12,4 +12,7 @@ export default defineConfig({
     __SERVER_PORT__: process.env.SERVER_PORT || 3001,
   },
   plugins: [react()],
+  build: {
+    
+  }
 });
